+++ tests/test_ux_features.py
@@ -344,11 +344,11 @@ class TestCombinedFeatures:
 class TestVersionUpdated:
     """Test that version is correct"""
 
-    def test_version_is_3_5_0(self):
-        """Test that version is 3.5.0"""
+    def test_version_is_3_5_1(self):
+        """Test that version is 3.5.1"""
         from cja_auto_sdr.generator import __version__
 
-        assert __version__ == "3.5.0"
+        assert __version__ == "3.5.1"
 
 
 class TestFormatAutoDetection: